+++ src/main/java/com/goobercorp/gooberlib/interfaces/WidgetProvider.java
@@ -5,5 +5,5 @@ import net.minecraft.client.gui.widget.ClickableWidget;
 
 public interface WidgetProvider<T extends Option<T>> {
 	// x and y are secretly always 0. shh, don't tell anyone!
-	ClickableWidget makeWidget(T theOption, int x, int y, double width, double height);
+	ClickableWidget makeWidget(T theOption, int x, int y, int width, int height);
 }
